:bug: Fix circular dependency in lisk-cryptography
