+++ packages/lisk-cryptography/src/convert.js
@@ -13,35 +13,10 @@
  *
  */
 import querystring from 'querystring';
-import bignum from 'browserify-bignum';
 import reverseBuffer from 'buffer-reverse';
 import ed2curve from 'ed2curve';
 import hash from './hash';
-
-export const bigNumberToBuffer = (bignumber, size) =>
-	bignum(bignumber).toBuffer({ size });
-
-export const bufferToBigNumberString = bigNumberBuffer =>
-	bignum.fromBuffer(bigNumberBuffer).toString();
-
-export const bufferToHex = buffer => Buffer.from(buffer).toString('hex');
-
-const hexRegex = /^[0-9a-f]+/i;
-export const hexToBuffer = (hex, argumentName = 'Argument') => {
-	if (typeof hex !== 'string') {
-		throw new TypeError(`${argumentName} must be a string.`);
-	}
-	const matchedHex = (hex.match(hexRegex) || [])[0];
-	if (!matchedHex || matchedHex.length !== hex.length) {
-		throw new TypeError(`${argumentName} must be a valid hex string.`);
-	}
-	if (matchedHex.length % 2 !== 0) {
-		throw new TypeError(
-			`${argumentName} must have a valid length of hex string.`,
-		);
-	}
-	return Buffer.from(matchedHex, 'hex');
-};
+import { bufferToBigNumberString } from './buffer';
 
 export const getFirstEightBytesReversed = publicKeyBytes =>
 	reverseBuffer(Buffer.from(publicKeyBytes).slice(0, 8));
