+++ packages/lisk-cryptography/src/index.js
@@ -12,10 +12,19 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
+import * as buffer from './buffer';
 import * as convert from './convert';
 import * as encrypt from './encrypt';
 import hash from './hash';
 import * as keys from './keys';
 import * as sign from './sign';
 
-export default Object.assign({}, convert, encrypt, { hash }, keys, sign);
+export default Object.assign(
+	{},
+	buffer,
+	convert,
+	encrypt,
+	{ hash },
+	keys,
+	sign,
+);
