+++ packages/lisk-cryptography/src/sign.js
@@ -16,7 +16,7 @@ import nacl from 'tweetnacl';
 import { encode as encodeVarInt } from 'varuint-bitcoin';
 import { SIGNED_MESSAGE_PREFIX } from 'lisk-constants';
 import hash from './hash';
-import { hexToBuffer, bufferToHex } from './convert';
+import { hexToBuffer, bufferToHex } from './buffer';
 import { getPrivateAndPublicKeyBytesFromPassphrase } from './keys';
 
 const createHeader = text => `-----${text}-----`;
